Removing RabbitMQ and using NSQ instead...
Rabbit MQ is a bit complicated for this use, and NSQ is already packaged as a Docker image. Now, we are letting Docker pull the image down and start the nsqlookupd service and the nsqd service. We also now have a constant [dockerHostIP] that finds and stores the HostIP of docker0 in a local variable until our provisioning script is finished running.